Adding MSSql database reference

+++ src/main/java/com/microsoft/egh/model/ClientDemographicRepository.java
@@ -3,7 +3,7 @@ package com.microsoft.egh.model;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface ClientDemographicRepository extends JpaRepository<ClientDemographic, Long> {
+public interface ClientDemographicRepository extends JpaRepository<ClientDemographic, Integer> {
     List<ClientDemographic> findByLastNameStartsWithIgnoreCase(String lastName);
 }
